[fix][doc] 소리감지, 라이트 시스템 수정
