--- Assets/Scripts/Car/HeadLightController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 382e1a366f391bb42aff3b2f8fe0030f
