+++ Assets/Scripts/Title/TitleButton.cs
@@ -52,7 +52,6 @@ public class TitleButton : MonoBehaviour
     IEnumerator CoFadeOut()
     {
         yield return new WaitForSeconds(0.1f);
-        float fadeAlp = 1;
         Color fadeCol = fadeImage.color;
         while (fadeCol.a > 0f)
         {
